chore(sdk-coin-etc): refactor etc to its own module
BG-48149
