+++ modules/account-lib/package.json
@@ -31,12 +31,12 @@
     "@bitgo/blake2b": "^3.0.2",
     "@bitgo/bls": "^0.1.0",
     "@bitgo/bls-dkg": "^1.0.2",
-    "@bitgo/ethereumjs-utils-old": "npm:ethereumjs-util@5.2.0",
     "@bitgo/sdk-coin-algo": "^1.0.0",
     "@bitgo/sdk-coin-avaxc": "^1.0.0",
     "@bitgo/sdk-coin-avaxp": "^1.0.0",
     "@bitgo/sdk-coin-celo": "^1.0.0",
     "@bitgo/sdk-coin-cspr": "^1.0.0",
+    "@bitgo/sdk-coin-etc": "^1.0.0",
     "@bitgo/sdk-coin-eth": "^1.0.0",
     "@bitgo/sdk-coin-rbtc": "^1.0.0",
     "@bitgo/sdk-coin-sol": "^1.0.0",
@@ -59,7 +59,6 @@
     "bn.js": "^5.2.1",
     "bs58": "^4.0.1",
     "ethereumjs-abi": "^0.6.5",
-    "ethereumjs-util": "6.2.1",
     "hi-base32": "^0.5.1",
     "joi": "^17.4.0",
     "js-sha256": "^0.9.0",
