+++ modules/bitgo/package.json
@@ -63,6 +63,7 @@
     "@bitgo/sdk-coin-cspr": "^1.0.0",
     "@bitgo/sdk-coin-dash": "^1.0.0",
     "@bitgo/sdk-coin-doge": "^1.0.0",
+    "@bitgo/sdk-coin-etc": "^1.0.0",
     "@bitgo/sdk-coin-eth": "^1.0.0",
     "@bitgo/sdk-coin-ltc": "^1.0.0",
     "@bitgo/sdk-coin-rbtc": "^1.0.0",
