+++ modules/sdk-coin-etc/.eslintignore
@@ -1,0 +1,5 @@
+node_modules
+.idea
+public
+dist
+
