+++ modules/sdk-coin-etc/.prettierignore
@@ -1,0 +1,2 @@
+.nyc_output/
+dist/
