+++ modules/sdk-coin-etc/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib';
+export * from './etc';
+export * from './tetc';
