+++ modules/sdk-coin-etc/src/lib/index.ts
