+++ modules/sdk-coin-etc/src/lib/utils.ts
