+++ modules/sdk-coin-etc/src/tetc.ts
